+++ src/uncms/jinja2_environment/pages.py
@@ -1,7 +1,33 @@
 import jinja2
 from django.template.loader import render_to_string
+from django.utils.safestring import mark_safe
 
 import uncms.pages.templatetags._common as pages_common
+from uncms.conf import defaults
+
+
+@jinja2.pass_context
+def get_breadcrumbs(context, *, breadcrumb_list=None, auto_extend=True, extend_with=None):
+    return pages_common.get_breadcrumbs_obj(
+        context,
+        breadcrumb_list=breadcrumb_list,
+        auto_extend=auto_extend,
+        extend_with=extend_with,
+    )
+
+
+@jinja2.pass_context
+def render_breadcrumbs(context, breadcrumbs_obj=None, *, auto_extend=True, show_tail=None, class_prefix='breadcrumbs'):
+    return mark_safe(render_to_string(
+        defaults.BREADCRUMBS_TEMPLATE.format(extension='jinja2'),
+        pages_common.get_breadcrumbs_context(
+            context,
+            breadcrumbs_obj=breadcrumbs_obj,
+            auto_extend=auto_extend,
+            show_tail=show_tail,
+            class_prefix=class_prefix,
+        ),
+    ))
 
 
 @jinja2.pass_context
@@ -26,6 +52,7 @@ PAGES_GLOBALS = {
     ]
 }
 
-
+PAGES_GLOBALS['get_breadcrumbs'] = get_breadcrumbs
+PAGES_GLOBALS['render_breadcrumbs'] = render_breadcrumbs
 PAGES_GLOBALS['render_title'] = render_title
 PAGES_GLOBALS['get_page_url'] = pages_common.get_page_url