+++ tests/pages/test_templates.py
@@ -1,3 +1,4 @@
+import pkg_resources
 import pytest
 from django.template.loader import render_to_string
 from django.test import RequestFactory
@@ -7,8 +8,31 @@ from tests.pages.factories import PageFactory
 from uncms.pages.middleware import RequestPageManager
 
 
+@pytest.mark.parametrize('django_template, jinja2_template', [
+    ('pages/templates/pages/breadcrumbs.html', 'pages/jinja2/pages/breadcrumbs.jinja2')
+])
+def test_django_and_jinja_templates_are_identical(django_template, jinja2_template):
+    """
+    To reduce the maintenance burden of having both Jinja2 and Django
+    templates, we should make sure that they are byte-for-byte identical.
+    We have to go to minor extra efforts elsewhere to ensure that they *can*
+    be (e.g. we track indexes in Python, rather than using "loop.index" or
+    "forloop.counter" in the templates). This is worthwhile.
+    """
+    jinja2_path = pkg_resources.resource_filename('uncms', jinja2_template)
+    django_path = pkg_resources.resource_filename('uncms', django_template)
+
+    with open(jinja2_path, encoding='utf-8') as fd:
+        jinja2_template_code = fd.read()
+
+    with open(django_path, encoding='utf-8') as fd:
+        django_template_code = fd.read()
+
+    assert jinja2_template_code == django_template_code
+
+
 @pytest.mark.django_db
-def test_head_meta_templates_are_identical(use_jinja2):
+def test_head_meta_templates_render_identically(use_jinja2):
     """
     Ensure that pages/head_meta.html and pages/head_meta.jinja2 are identical
     (minus any trailing newlines) to ensure parity.